requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
openai==2.8.1
openai-chatkit==1.3.0
python-dotenv==1.2.1
pydantic-settings==2.12.0
